requirements.txt
mcp>=1.2.0
httpx>=0.27.0
starlette>=0.27.0
python-dotenv>=1.0.0
requests>=2.31.0  # For LangGraph HTTP mode1
starlette>=0.27.0
